requirements.txt
torch
torchvision
transformers
langchain
openai
spacy
nltk
streamlit
fastapi
uvicorn
pillow